--- a/fmt_ym.go
+++ b/fmt_ym.go
@@ -8,7 +8,7 @@ import (
 
 //nolint:cyclop,gocognit
 func fmtYearMonthGregorian(locale language.Tag, digits digits, opts Options) func(y int, m time.Month) string {
-	var month func(time.Month) string
+	var month func(int) string
 
 	lang, script, region := locale.Raw()
 	year := fmtYear(digits, opts.Year)
@@ -424,12 +424,12 @@ func fmtYearMonthGregorian(locale language.Tag, digits digits, opts Options) fun
 
 	if layout == layoutMonthYear {
 		return func(y int, m time.Month) string {
-			return prefix + month(m) + middle + year(y) + suffix
+			return prefix + month(int(m)) + middle + year(y) + suffix
 		}
 	}
 
 	return func(y int, m time.Month) string {
-		return prefix + year(y) + middle + month(m) + suffix
+		return prefix + year(y) + middle + month(int(m)) + suffix
 	}
 }
 
@@ -440,7 +440,7 @@ func fmtYearMonthBuddhist(locale language.Tag, digits digits, opts Options) func
 		month := fmtMonth(digits, opts.Month)
 
 		return func(y int, m time.Month) string {
-			return month(m) + "/" + year(y)
+			return month(int(m)) + "/" + year(y)
 		}
 	}
 
@@ -448,7 +448,7 @@ func fmtYearMonthBuddhist(locale language.Tag, digits digits, opts Options) func
 	month := fmtMonth(digits, Month2Digit)
 
 	return func(y int, m time.Month) string {
-		return prefix + year(y) + "-" + month(m)
+		return prefix + year(y) + "-" + month(int(m))
 	}
 }
 
@@ -467,7 +467,7 @@ func fmtYearMonthPersian(locale language.Tag, digits digits, opts Options) func(
 		// year=2-digit,month=numeric,out=١٠/٠٢
 		// year=2-digit,month=2-digit,out=١٠/٠٢
 		return func(y int, m time.Month) string {
-			return month(m) + "/" + year(y)
+			return month(int(m)) + "/" + year(y)
 		}
 	case fa:
 		separator = "/"
@@ -489,6 +489,6 @@ func fmtYearMonthPersian(locale language.Tag, digits digits, opts Options) func(
 	}
 
 	return func(y int, m time.Month) string {
-		return prefix + year(y) + separator + month(m)
+		return prefix + year(y) + separator + month(int(m))
 	}
 }