--- a/fmt_ymd.go
+++ b/fmt_ymd.go
@@ -12,7 +12,7 @@ func fmtYearMonthDayGregorian(
 	digits digits,
 	opts Options,
 ) func(y int, m time.Month, d int) string {
-	var month func(time.Month) string
+	var month func(int) string
 
 	lang, script, region := locale.Raw()
 	year := fmtYear(digits, opts.Year)
@@ -834,7 +834,7 @@ func fmtYearMonthDayGregorian(
 		day := fmtDay(digits, opts.Day)
 
 		return func(y int, m time.Month, d int) string {
-			return day(d) + "/" + month(m) + " " + year(y)
+			return day(d) + "/" + month(int(m)) + " " + year(y)
 		}
 	case ko:
 		// year=numeric,month=numeric,day=numeric,out=2024. 1. 2.
@@ -1299,7 +1299,7 @@ func fmtYearMonthDayGregorian(
 				day := fmtDay(digits, opts.Day)
 
 				return func(y int, m time.Month, d int) string {
-					return year(y) + "年" + month(m) + "月" + day(d) + "日"
+					return year(y) + "年" + month(int(m)) + "月" + day(d) + "日"
 				}
 			}
 
@@ -1357,19 +1357,19 @@ func fmtYearMonthDayGregorian(
 	switch layout {
 	default: // layoutYearMonthDay
 		return func(y int, m time.Month, d int) string {
-			return prefix + year(y) + separator + month(m) + separator + day(d) + suffix
+			return prefix + year(y) + separator + month(int(m)) + separator + day(d) + suffix
 		}
 	case layoutDayMonthYear:
 		return func(y int, m time.Month, d int) string {
-			return day(d) + separator + month(m) + separator + year(y) + suffix
+			return day(d) + separator + month(int(m)) + separator + year(y) + suffix
 		}
 	case layoutMonthDayYear:
 		return func(y int, m time.Month, d int) string {
-			return month(m) + separator + day(d) + separator + year(y) + suffix
+			return month(int(m)) + separator + day(d) + separator + year(y) + suffix
 		}
 	case layoutYearDayMonth:
 		return func(y int, m time.Month, d int) string {
-			return year(y) + separator + day(d) + separator + month(m) + suffix
+			return year(y) + separator + day(d) + separator + month(int(m)) + suffix
 		}
 	}
 }
@@ -1439,12 +1439,12 @@ func fmtYearMonthDayPersian(
 
 	if layout == layoutDayMonthYear {
 		return func(y int, m time.Month, d int) string {
-			return day(d) + "/" + month(m) + "/" + year(y)
+			return day(d) + "/" + month(int(m)) + "/" + year(y)
 		}
 	}
 
 	return func(y int, m time.Month, d int) string {
-		return prefix + year(y) + separator + month(m) + separator + day(d)
+		return prefix + year(y) + separator + month(int(m)) + separator + day(d)
 	}
 }
 
@@ -1467,6 +1467,6 @@ func fmtYearMonthDayBuddhist(
 	// year=2-digit,month=2-digit,day=numeric,out=2/01/24
 	// year=2-digit,month=2-digit,day=2-digit,out=02/01/24
 	return func(y int, m time.Month, d int) string {
-		return day(d) + "/" + month(m) + "/" + year(y)
+		return day(d) + "/" + month(int(m)) + "/" + year(y)
 	}
 }