I raised a separate issue to speed up the formatting of the Latin numbers #79 

```console
✗ benchstat a.txt b.txt          
goos: darwin
goarch: arm64
pkg: go.expect.digital/intl
cpu: Apple M2 Max
                               │    a.txt    │                b.txt                │
                               │   sec/op    │   sec/op     vs base                │
FmtEraYearMonthDayGregorian-12   115.7n ± 1%   125.1n ± 1%   +8.13% (p=0.000 n=10)
NewDateTime/fa-IR-12             28.68n ± 0%   27.82n ± 0%   -2.98% (p=0.000 n=10)
NewDateTime/lv-LV-12             22.27n ± 0%   22.51n ± 1%   +1.08% (p=0.000 n=10)
NewDateTime/dz-BT-12             22.13n ± 0%   21.68n ± 1%   -2.06% (p=0.002 n=10)
DateTime_Format/fa-IR-12         684.0n ± 2%   596.7n ± 1%  -12.76% (p=0.000 n=10)
DateTime_Format/lv-LV-12         346.6n ± 2%   381.9n ± 2%  +10.20% (p=0.000 n=10)
DateTime_Format/dz-BT-12         545.8n ± 2%   398.3n ± 0%  -27.01% (p=0.000 n=10)
geomean                          111.3n        106.4n        -4.42%

                               │    a.txt     │                b.txt                 │
                               │     B/op     │    B/op     vs base                  │
FmtEraYearMonthDayGregorian-12   32.00 ± 0%     44.00 ± 0%  +37.50% (p=0.000 n=10)
NewDateTime/fa-IR-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12         192.0 ± 0%     152.0 ± 0%  -20.83% (p=0.000 n=10)
DateTime_Format/lv-LV-12         72.00 ± 0%     93.00 ± 0%  +29.17% (p=0.000 n=10)
DateTime_Format/dz-BT-12         208.0 ± 0%     152.0 ± 0%  -26.92% (p=0.000 n=10)
geomean                                     ²                +0.39%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                               │    a.txt     │                b.txt                 │
                               │  allocs/op   │ allocs/op   vs base                  │
FmtEraYearMonthDayGregorian-12   4.000 ± 0%     4.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/fa-IR-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12         15.00 ± 0%     11.00 ± 0%  -26.67% (p=0.000 n=10)
DateTime_Format/lv-LV-12         12.00 ± 0%     12.00 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12         20.00 ± 0%     11.00 ± 0%  -45.00% (p=0.000 n=10)
geomean                                     ²               -12.17%                ²
¹ all samples are equal
```